+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+* Implemented local notifications.
+    - Added `DIALOG.notification`.
+    - **Breaking** Added to the view API.
+
+* Fix file dialog tasks never responding if the view-process crashes while the dialog is showing.
 * Optimized `Txt`, now it is 24 bytes.
     - **Breaking** `Txt::from_string` is no longer const.